+++ eslint.config.mjs
@@ -1,0 +1,1 @@
+export { default } from "@topo/eslint-config";